+++ CHANGELOG.md
@@ -5,9 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html)
 
+## [0.1.44] - 2021-09-27
+
+-- fix cluster optional forcerestart
+
+### Changed
+-  api/src/main/java/com/homeaway/datapullclient/process/DataPullTask.java
+
 ## [0.1.43] - 2021-09-22
 
--- adding optional forcerestart in datapull
+-- adding cluster optional forcerestart in datapull
 
 ### Changed
 -  api/src/main/java/com/homeaway/datapullclient/config/EMRProperties.java